toss refactor, get rid of json guildfiles, rotate logfile

+++ fluff/service/RolebanService.py
@@ -14,7 +14,7 @@ from service.ConfigService import ConfigService
 from service.NotificationService import NotificationService
 
 RULEPUSH_CHANNEL_NAME_PATTERN = re.compile(r"^rulepush(\d+)$")
-TOSS_CHANNEL_NAME_PATTERN = re.compile(r"^toss(\d+)$")
+TOSS_CHANNEL_NAME_PATTERN = re.compile(r"^muted(\d+)$")
 
 class RolebanService:
     """A service that is responsible for handling all things roleban (toss/untoss, rulepush, etc)"""
@@ -33,7 +33,7 @@ class RolebanService:
 
         return None
 
-    async def roleban_users(self, ctx: commands.Context, members: list[discord.Member], roleban_type: RolebanType, remove_roles: bool = True) -> RolebanSession | None:
+    async def roleban_users(self, ctx: commands.Context, members: list[discord.Member], roleban_type: RolebanType, remove_roles: bool = True, session_to_use: RolebanSession = None) -> RolebanSession | None:
         """Rolebans a user by creating a new roleban channel, removing the users roles (if remove_roles = true), and adding the roleban role
         Returns: an instance of RolebanSession if the user was successfully rolebanned, otherwise None, meaning an error occurred and the user was not rolebanned"""
 
@@ -53,9 +53,15 @@ class RolebanService:
                 )
                 return None
 
-        channel: TextChannel | None = await self.create_roleban_channel(ctx.guild, roleban_type)
+        channel: TextChannel | None = None
+        if session_to_use:
+            channel = await self.get_channel(session_to_use.channel_id)
+        else:
+            channel = await self.create_roleban_channel(ctx.guild, roleban_type)
+
         if channel is None:
-            await ctx.reply("Failed to create the roleban channel.", mention_author=False)
+            channel_action_type: str = "lookup" if session_to_use else "create"
+            await ctx.reply(f"Failed to {channel_action_type} the roleban channel.", mention_author=False)
             return None
 
         roleban_role = self.bot.pull_role(ctx.guild, self.config_service.get_server_config(ctx.guild.id, "toss", "tossrole"))
@@ -72,16 +78,18 @@ class RolebanService:
                 user_ids_to_unassignable_roles[member.id] = list()
 
         try:
-            session = await self.roleban_repo.create_session(
-                ctx.guild.id,
-                user_ids_to_all_roles,
-                channel.id,
-                ctx.author.id,
-                roleban_type
+            session = await self.roleban_repo.create_or_add_to_session(
+                server_id=ctx.guild.id,
+                user_ids_to_roles=user_ids_to_all_roles,
+                channel_id=channel.id,
+                rolebanned_by=ctx.author.id,
+                roleban_type=roleban_type,
+                session_id_to_add=session_to_use.id if session_to_use else None
             )
         except sqlite3.Error as err:
             self.bot.log.error(f"Error creating roleban session in server {ctx.guild.id}: {err}")
-            await self.delete_roleban_channel(channel, "roleban setup failed")
+            if not session_to_use: #if an existing roleban channel already exists with users, and we failed to add a new user to it, then dont delete the existing session
+                await self.delete_roleban_channel(channel=channel, reason="roleban setup failed", session_id=None, delete_session=False)
             await ctx.reply("Error while creating the roleban session.", mention_author=False)
             return None
 
@@ -102,12 +110,13 @@ class RolebanService:
             except Exception as err:
                 self.bot.log.error(f"Error sending notification embed for roleban: {err}")
 
-        #some portion of the users either couldnt be assigned the correct role, or couldnt view the channel, so we need to unroleban everyone
+        #some portion of the users either could not be assigned the correct role, or could not view the channel, so we need to unroleban everyone
         if not all_successful:
             for member in members:
                 await self.unroleban_user(ctx, session, member.id, channel)
 
-            await self.delete_roleban_channel(channel, "roleban setup failed", session.id)
+            if not session_to_use:  # if an existing roleban channel already exists with users, and we failed to add a new user to it, then dont delete the existing session
+                await self.delete_roleban_channel(channel=channel, reason="roleban setup failed", session_id=session.id, delete_session=True)
             await ctx.reply("Error attempting to set roleban permissions for user. Roleban cancelled.", mention_author=False)
             return None
 
@@ -115,16 +124,15 @@ class RolebanService:
 
     async def unroleban_user(self, ctx: commands.Context, session: RolebanSession, user_id: int, channel: discord.TextChannel | None = None) -> bool:
         """Releases a user from their roleban, restores roles (if the member is present), and sends a message in the notification channel
-        Returns: True if this codepath successfully unrolebanned the user, false we were unable to unroleban the user"""
+        Returns: True if this codepath successfully unrolebanned the user, false if we were unable to unroleban the user"""
         if user_id not in {u.user_id for u in session.users}:
             return False
 
         member = ctx.guild.get_member(user_id)
         try:
             role_ids = await self.roleban_repo.get_role_ids(session.id, user_id)
-            await self.roleban_repo.remove_user_from_session(session.id, user_id)
         except sqlite3.Error as err:
-            self.bot.log.error(f"Error releasing roleban session for user ID user_id {user_id} and session ID {session.id}: {err}")
+            self.bot.log.error(f"Error fetching rolebanned user roles for user ID {user_id} and session ID {session.id}: {err}")
             return False
 
         restored_roles, failed_roles = [], []
@@ -134,13 +142,27 @@ class RolebanService:
                 roles = [role for role in roles if role is not None]
                 assignable = [role for role in roles if role.is_assignable()]
                 unassignable = [role for role in roles if not role.is_assignable()]
-                restored_roles, failed_roles = await self.replace_roles(member, assignable, unassignable, f"User unrolebanned by {ctx.author} ({ctx.author.id})")
+                try:
+                    restored_roles, failed_roles = await self.replace_roles(member, assignable, unassignable, f"User unrolebanned by {ctx.author} ({ctx.author.id})")
+                except Exception as err:
+                    self.bot.log.error(f"Error re-assigning roles to rolebanned user: {err}")
+                    return False
+            else:
+                removed_roleban_role: bool = await self.remove_roleban_role(member)
+                if not removed_roleban_role:
+                    self.bot.log.error(f"Could not remove roleban role from user: {member.id}")
+                    return False
             if channel is not None:
                 try:
                     await channel.set_permissions(member, read_messages=False)
                 except Exception as err:
                     pass
 
+        try:
+            await self.roleban_repo.remove_user_from_session(session.id, user_id)
+        except sqlite3.Error as err:
+            self.bot.log.error(f"Error deleting roleban session for user ID {user_id} and session ID {session.id}: {err}")
+            return False
 
         embed = stock_embed(self.bot)
         embed.color = discord.Color.green()
@@ -152,6 +174,8 @@ class RolebanService:
 
         if member is not None and member.id == ctx.author.id:
             embed.description = f"{member.mention} was un{session.type.value}ed automatically [`#{ctx.channel.name}`]"
+        elif member is not None:
+            embed.description = f"{member.mention} was un{session.type.value}ed by {ctx.author.mention} [`#{ctx.channel.name}`]"
         else:
             embed.description = f"user ID {user_id} was un{session.type.value}ed by {ctx.author.mention} [`#{ctx.channel.name}`]"
 
@@ -222,12 +246,13 @@ class RolebanService:
             overwrites=overwrites
         )
 
-    async def delete_roleban_channel(self, channel: discord.TextChannel, reason: str, session_id: int | None = None) -> bool:
+    async def delete_roleban_channel(self, channel: discord.TextChannel, reason: str, session_id: int | None = None, delete_session: bool = False) -> bool:
         """Deletes a roleban channel if no active users are still rolebanned in that channel
            Args:
                   channel: discord.TextChannel to delete
                   reason: reason for deleting this channel
                   session_id: optional ID. If passed, validates that no active users are still rolebanned in that session
+                  delete_session: boolean whether to delete this session. skipped if session_id is None
            Returns: a boolean representing whether the channel was deleted or not"""
         if not RULEPUSH_CHANNEL_NAME_PATTERN.match(channel.name) and not TOSS_CHANNEL_NAME_PATTERN.match(channel.name):
             return False
@@ -238,7 +263,8 @@ class RolebanService:
                 if len(active_session_users) > 0:
                     return False
 
-                await self.delete_session(session_id)
+                if delete_session:
+                    await self.delete_session(session_id)
 
             await channel.delete(reason=reason)
             return True
@@ -275,14 +301,62 @@ class RolebanService:
         except sqlite3.Error as err:
             self.bot.log.error(f"Error updating session start time for session ID {session_id}: {err}")
 
-    async def reactivate_user_session(self, session_id: int, user_id: int, channel_id: int) -> int:
-        """Reactivates a users session by setting the status to ACTIVE and updating the channel ID"""
-        try:
-            return await self.roleban_repo.reactivate_user_session(session_id, user_id, channel_id)
-        except sqlite3.Error as err:
-            self.bot.log.error(f"Error reactivating session for session ID {session_id} and user ID {user_id}: {err}")
+    async def reactivate_user_session(self, guild: discord.Guild, member: discord.Member, roleban_type: RolebanType) -> tuple[discord.TextChannel, int, bool] | None:
+        """Reactivates a users session by re-creating the roleban channel (if necessary), and re-assigning the roleban role
+        Returns: a tuple of [discord.TextChannel, session ID, bool] where the channel represents the actual channel this roleban is taking place,
+        an int, representing the session ID in the database for this roleban, and a Boolean representing whether this channel was created
+        from this method call, in which case the Boolean is true, or the channel already existed, in which case this Boolean is false. If this
+        method fails, None is returned, indicating we did not successfully re-activate/roleban the user."""
 
-        return 0
+        try:
+            session: RolebanSession | None = await self.get_roleban_session_by_user(guild.id, member.id)
+        except sqlite3.Error as err:
+            self.bot.log.error(f"Error fetching session info when attempting to re-activate user session for user: {member.id}: {err}")
+            return None
+
+        if session is None or session.type is not roleban_type or session.users is None or len(session.users) <= 0:
+            return None
+
+        # reuse the old channel if it still exists, otherwise recreate it
+        channel: discord.TextChannel = self.bot.get_channel(session.channel_id) if session.channel_id else None
+        channel_recreated = False
+        if channel is None:
+            channel = await self.create_roleban_channel(guild, roleban_type)
+            channel_recreated = True
+
+            # channel creation failed
+            if channel is None:
+                await self.send_roleban_resume_failure_notification(member,roleban_type, "I could not create a new channel")
+                return None
+
+        try:
+            await channel.set_permissions(member, read_messages=True)
+        except (discord.Forbidden, discord.HTTPException) as err:
+            self.bot.log.error(f"Error restoring channel permissions on rejoin in server {guild.id}: {err}")
+            await self.send_roleban_resume_failure_notification(member, roleban_type,"I could not assign the correct channel permissions for the user")
+            return None
+
+        successful = await self.assign_roleban_role(member, "User was automatically rolebanned on server join.")
+        if not successful:
+            await self.send_roleban_resume_failure_notification(member,roleban_type, "I could not assign the roleban role to the user")
+            return None
+
+        try:
+            await self.roleban_repo.reactivate_user_session(session.id, member.id, channel.id)
+        except sqlite3.Error as err:
+            self.bot.log.error(f"Error reactivating session for session ID {session.id} and user ID {member.id}: {err}")
+            await self.send_roleban_resume_failure_notification(member,roleban_type, f"a database error prevented the session from being reactivated. The channel is {channel.mention}")
+            return None
+
+        embed = stock_embed(self.bot)
+        embed.title = f"🔁 {roleban_type.value.capitalize()} Resumed"
+        embed.color = discord.Color.orange()
+        embed.description = (
+            f"{member.mention} ({member.id}) rejoined while {roleban_type.value.lower()}ed. Continuing in {channel.mention}..."
+        )
+        await self.notification_service.send_notification(member.guild, embed)
+
+        return channel, session.id, channel_recreated
 
     async def replace_roles(self, member: discord.Member, assignable_roles: list[discord.Role],
                             unassignable_roles: list[discord.Role], reason: str) -> tuple[list, list]:
@@ -312,6 +386,20 @@ class RolebanService:
 
         return True
 
+    async def remove_roleban_role(self, member: discord.Member) -> bool:
+        """removes the roleban role from the user, if it exists"""
+        try:
+            roleban_role = self.bot.pull_role(member.guild, self.config_service.get_server_config(member.guild.id, "toss", "tossrole"))
+            if roleban_role is None:
+                return False
+            await member.remove_roles(roleban_role)
+        except Exception as e:
+            self.bot.log.error(f"Failed to remove roleban role to user: {member.id}")
+            return False
+
+        return True
+
+
 
     async def get_non_rolebanned_user_roles(self, guild: discord.Guild, member: discord.Member, roleban_role: discord.Role) -> tuple[list[discord.Role], list[discord.Role]]:
         """Fetches all roles (excluding the guild default role and the roleban role) that belong to a user, and unassignable roles,
@@ -322,7 +410,7 @@ class RolebanService:
         return all_roles, unassignable_roles
 
     def get_next_channel_name(self, guild: discord.Guild, roleban_type: RolebanType) -> str:
-        """Picks the lowest roleban-N name not currently in use in this guild"""
+        """Picks the lowest rolebanN name not currently in use in this guild"""
         used = set()
         for channel in guild.channels:
             match = None
@@ -336,6 +424,10 @@ class RolebanService:
         number = 1
         while number in used:
             number += 1
+
+        if roleban_type == RolebanType.TOSS:
+            return f"muted{number}"
+
         return f"{roleban_type.value}{number}"
 
     def create_notification_embed(self, ctx: commands.Context, member: discord.Member,
@@ -373,6 +465,14 @@ class RolebanService:
 
         return notify_embed
 
+    async def send_roleban_resume_failure_notification(self, member: discord.Member, roleban_type: RolebanType, reason: str):
+        """Sends a notification to the modlog channel, detailing why we were unable to resume a previous roleban"""
+        embed = stock_embed(self.bot)
+        embed.title = f"⚠️ {roleban_type.value.capitalize()} Resume Failed"
+        embed.color = discord.Color.red()
+        embed.description = f"{member.mention} rejoined with a pending {roleban_type.value.lower()} roleban, but {reason}."
+        return await self.notification_service.send_notification(member.guild, embed)
+
     def format_role_list(self, roles: list[discord.Role]) -> str:
         if not roles:
             return "None"
@@ -380,3 +480,14 @@ class RolebanService:
         if len(formatted) > 1024:  # embed field value limit
             formatted = formatted[:1021] + "..."
         return formatted
+
+    async def get_channel(self, channel_id: int) -> TextChannel | None:
+        channel = self.bot.get_channel(channel_id)
+        if channel is None:
+            try:
+                channel = await self.bot.fetch_channel(channel_id)
+            except discord.HTTPException as e:
+                self.bot.log.error(f"Failed to fetch channel {channel_id}: {e}")
+                return None
+
+        return channel
